FXML and UML Updated some code and tests.

+++ src/main/java/ucf/assignments/InventoryItem.java
@@ -5,10 +5,10 @@ public class InventoryItem {
     //Create attributes for the name, serial number and value of an item.
     String name;
     String serialNumber;
-    int value;
+    double value;
 
     //Create a constructor.
-    public InventoryItem(String name, String serialNumber, int value) {
+    public InventoryItem(String name, String serialNumber, double value) {
         //Initialize the attributes of the object with the information passed as parameters.
         this.name = name;
         this.serialNumber = serialNumber;
@@ -20,14 +20,14 @@ public class InventoryItem {
     //Create a function to set the serial number of an item.
     public void setSerialNumber(String serialNumber) {this.serialNumber = serialNumber;}
     //Create a function to set the value of an item.
-    public void setValue(int value) {this.value = value;}
+    public void setValue(double value) {this.value = value;}
 
     //Create a function that returns the name of an item.
     public String getName() {return name;}
     //Create a function that returns the serial number of item.
     public String getSerialNumber() {return serialNumber;}
     //Create a function that returns the value of item.
-    public int getValue() {return value;}
+    public double getValue() {return value;}
 
     //Create a to string function that returns the information of an item in TSV format.
     @Override
